modules/feature-engineering/requirements.txt
numpy==1.26.2
pandas==2.1.3
xarray==2023.11.0
rasterio==1.3.9
geopandas==0.14.1
scipy==1.11.4
scikit-learn==1.3.2
netCDF4==1.6.5
python-dotenv==1.0.0
